Removed C-style casts which break for 64-bit targets.

+++ trunk/editor/editorglwidget.cpp
@@ -85,7 +85,7 @@ void EditorGLWidget::setMap(hex::gamemap *map) {
 
 void EditorGLWidget::setCamera(hex::camera *camera) {
 	camera_ = camera;
-	std::cerr << "setting camera_ to " << std::hex << (int)camera << std::endl;
+	std::cerr << "setting camera_ to " << std::hex << camera << std::endl;
 	camera_->set_dim(width(),height());
 	std::cerr << "setting dim to " << std::dec << width() << ", " << height() << std::endl;
 }
